remotehire_backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
python-multipart
pytz
spacy
pdfminer.six
python-docx 
django-cors-headers
gunicorn
requests
pillow
google-auth
google-auth-oauthlib
google-auth-httplib2
python-dotenv
google-genai
pypdf
psycopg2-binary
boto3
django-storages
openai
groq